hectre-be-app/pyproject.toml
[tool.poetry]
name = "hectre-be-app"
version = "0.1.0"
description = ""
authors = ["Manny Kwaning <ema526@g.harvard.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.110.0"
loguru = "^0.7.2"
SQLAlchemy = "^2.0.28"
boto3 = "^1.34.65"
botocore = "^1.34.65"
python-magic = "^0.4.27"

[tool.poetry.dev-dependencies]
pytest = "^8.1.1"
black = "^20.8b1"
flake8 = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"